requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
black==24.10.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.1.2
django-appconf==1.0.6
django-crispy-forms==2.3
django-filter==23.2
django-select2==8.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-nested-routers==0.94.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
qrcode==8.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
timedelta==2020.12.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.0
